+++ jest.config.js
@@ -25,7 +25,8 @@ module.exports = {
   },
 
   // The glob patterns Jest uses to detect test files
-  testMatch: ["**/test/**/?(*.)+(spec|test).js?(x)"],
+  testMatch: ["<rootDir>/test/**/*.spec.js"],
+  // testMatch: ["<rootDir>/test/**/commands.spec.js"],
 
   // The test environment that will be used for testing
   testEnvironment: "node",