+++ test-e2e/cypress/integration/mocks.js
@@ -2,6 +2,8 @@ describe("Mocks server responses", () => {
   const SELECTORS = {
     RESPONSE: "#response",
     RESPONSE_TIME: "#response-time",
+    LEGACY_RESPONSE: "#legacy-response",
+    LEGACY_RESPONSE_TIME: "#legacy-response-time",
   };
 
   describe("when started", () => {
@@ -22,14 +24,18 @@ describe("Mocks server responses", () => {
     });
   });
 
-  describe("when behavior is changed to custom", () => {
+  describe("when mock is changed to custom", () => {
     before(() => {
-      cy.mocksServerSetBehavior("custom");
+      cy.mocksSetMock("custom");
       cy.visit("/");
     });
 
     it("should display custom response", () => {
       cy.get(SELECTORS.RESPONSE).should("have.text", "custom-response");
+    });
+
+    it("should display standard legacy response", () => {
+      cy.get(SELECTORS.LEGACY_RESPONSE).should("have.text", "legacy-standard-response");
     });
 
     it("should load response fast", () => {
@@ -41,9 +47,32 @@ describe("Mocks server responses", () => {
     });
   });
 
+  describe("when behavior is changed to custom", () => {
+    before(() => {
+      cy.mocksSetBehavior("custom");
+      cy.visit("/");
+    });
+
+    it("should display custom response", () => {
+      cy.get(SELECTORS.RESPONSE).should("have.text", "custom-response");
+    });
+
+    it("should display custom legacy response", () => {
+      cy.get(SELECTORS.LEGACY_RESPONSE).should("have.text", "legacy-custom-response");
+    });
+
+    it("should load legacy response fast", () => {
+      cy.get(SELECTORS.LEGACY_RESPONSE_TIME).should(($div) => {
+        const text = $div.text();
+
+        expect(Number(text)).to.be.lessThan(1000);
+      });
+    });
+  });
+
   describe("when delay is changed", () => {
     before(() => {
-      cy.mocksServerSetDelay(1000);
+      cy.mocksSetDelay(1000);
       cy.visit("/");
     });
 
@@ -58,12 +87,20 @@ describe("Mocks server responses", () => {
         expect(Number(text)).to.be.greaterThan(1000);
       });
     });
+
+    it("should load legacy response with delay", () => {
+      cy.get(SELECTORS.LEGACY_RESPONSE_TIME).should(($div) => {
+        const text = $div.text();
+
+        expect(Number(text)).to.be.greaterThan(1000);
+      });
+    });
   });
 
   describe("when settings are changed", () => {
     before(() => {
-      cy.mocksServerSetSettings({
-        behavior: "standard",
+      cy.mocksSetSettings({
+        mock: "standard",
         delay: 0,
       });
       cy.visit("/");
@@ -79,6 +116,28 @@ describe("Mocks server responses", () => {
 
         expect(Number(text)).to.be.lessThan(1000);
       });
+    });
+  });
+
+  describe("when custom route variant is used", () => {
+    before(() => {
+      cy.mocksUseRouteVariant("response:custom");
+      cy.visit("/");
+    });
+
+    it("should display custom response", () => {
+      cy.get(SELECTORS.RESPONSE).should("have.text", "custom-response");
+    });
+  });
+
+  describe("when route variants are restored", () => {
+    before(() => {
+      cy.mocksRestoreRoutesVariants();
+      cy.visit("/");
+    });
+
+    it("should display standard response", () => {
+      cy.get(SELECTORS.RESPONSE).should("have.text", "standard-response");
     });
   });
 });