+++ test/register.spec.js
@@ -15,16 +15,28 @@ describe("register", () => {
   });
 
   describe("commands", () => {
-    it("should register mocksServerSetBehavior command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksServerSetBehavior")).toBe(true);
+    it("should register mocksSetMock command", () => {
+      expect(cypressMock.stubs.Commands.add.calledWith("mocksSetMock")).toBe(true);
     });
 
-    it("should register mocksServerSetDelay command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksServerSetDelay")).toBe(true);
+    it("should register mocksSetDelay command", () => {
+      expect(cypressMock.stubs.Commands.add.calledWith("mocksSetDelay")).toBe(true);
     });
 
-    it("should register mocksServerSetSettings command", () => {
-      expect(cypressMock.stubs.Commands.add.calledWith("mocksServerSetSettings")).toBe(true);
+    it("should register mocksSetSettings command", () => {
+      expect(cypressMock.stubs.Commands.add.calledWith("mocksSetSettings")).toBe(true);
+    });
+
+    it("should register mocksUseRouteVariant command", () => {
+      expect(cypressMock.stubs.Commands.add.calledWith("mocksUseRouteVariant")).toBe(true);
+    });
+
+    it("should register mocksRestoreRoutesVariants command", () => {
+      expect(cypressMock.stubs.Commands.add.calledWith("mocksRestoreRoutesVariants")).toBe(true);
+    });
+
+    it("should register mocksSetBehavior command", () => {
+      expect(cypressMock.stubs.Commands.add.calledWith("mocksSetBehavior")).toBe(true);
     });
   });
 });